Separated level=0 and levels>0 alignments around large icicles. JIRA: CXX-9368
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@78475 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/sra/data_loaders/bam/impl/bamloader_impl.hpp
@@ -73,7 +73,7 @@ public:
         }
 
     // estimated number of alignments
-    size_t GetAlignCount(void) const
+    Uint8 GetAlignCount(void) const
         {
             return m_AlignCount;
         }
@@ -114,7 +114,7 @@ public:
 protected:
     friend class CBamRefSeqInfo;
 
-    size_t m_AlignCount; // 0 - graph only
+    Uint8 m_AlignCount; // 0 - graph only
     TRange m_RefSeqRange;
     TSeqPos m_MaxRefSeqFrom;
 };
